% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_theories_2}
\alias{all_theories_2}
\title{Theory predictions: 2x2 games}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  ...
}
}
\source{
\url{http://madeitup.myself/}
}
\usage{
all_theories_2
}
\description{
Blabla!
}
\keyword{datasets}
