% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_theories_3}
\alias{all_theories_3}
\title{Predictions of all 3x3 theories}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  ...
}
}
\source{
\url{http://madeitup.myself/}
}
\usage{
all_theories_3
}
\description{
Blabla!
}
\keyword{datasets}
