% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vuong_table.R
\name{color_vuong_table}
\alias{color_vuong_table}
\title{Color a Vuong matrix in red/green
requires the xcolor package and /definecolor{darkred}{rgb}{0.55, 0.0, 0.0}
Red/green, but still kind of readable in black/white}
\usage{
color_vuong_table(vuong_table)
}
\arguments{
\item{vuong_table}{a Voung table like the output of vuong_matrix}
}
\value{
Latex code of a colored vuong table
}
\description{
Color a Vuong matrix in red/green
requires the xcolor package and /definecolor{darkred}{rgb}{0.55, 0.0, 0.0}
Red/green, but still kind of readable in black/white
}
\examples{
(missing)
}
