% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_two_action_games}
\alias{data_two_action_games}
\title{Data from the two-action games}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  \item{cells}{number of subjects who chose each action on each treatment}
  ...
}
}
\source{
\url{http://hdg.kuelpmann.org/}
}
\usage{
data_two_action_games
}
\description{
Columns are different treatments, rows are actions and the cells are the number of subjects who chose each action on each treatment
}
\keyword{datasets}
