% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{get_all_chi_sq}
\alias{get_all_chi_sq}
\title{Chi-Square Tests for multiple theories}
\usage{
get_all_chi_sq(data, predictions)
}
\arguments{
\item{data}{Data of the experiment (matrix of ints/data.frame): columns are observations of action 1,2,3..., rows are treatments.}

\item{predictions}{Model prediction for one theory: rows/cols as data}
}
\value{
Output of all chi square tests
}
\description{
Chi-Square Tests for multiple theories
}
