% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lh.R
\name{log_lh}
\alias{log_lh}
\title{Calculate log-likelihood for a theory}
\usage{
log_lh(model_prediction, data_experiment)
}
\arguments{
\item{model_prediction}{Model prediction for a theory (matrix of doubles/data.frame): columns are prediction of playing action 1,2,3..., rows are treatments}

\item{data_experiment}{Data of the experiment (matrix of ints/data.frame): columns are observations of action 1,2,3..., rows are treatments}
}
\value{
The log-likelihood of the theory for each treatment
}
\description{
Calculate log-likelihood for a theory
}
