% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oottest-package.R
\docType{package}
\name{oottest-package}
\alias{oottest}
\alias{oottest-package}
\title{oottest: Out-of-Treatment Testing}
\description{
Implements the out-of-treatment testing from Kuelpmann and Kuzmics (2020). TODO: need to add more things here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhilippKuelpmann/oottest}
  \item Report bugs at \url{https://github.com/PhilippKuelpmann/oottest/issues}
}

}
\author{
\strong{Maintainer}: Philipp Kuelpmann \email{philipp@kuelpmann.org} (\href{https://orcid.org/0000-0003-3757-7172}{ORCID})

Authors:
\itemize{
  \item Panagiota Charalampidou
}

}
\keyword{internal}
