% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictions_three_action_games}
\alias{predictions_three_action_games}
\title{Predictions for the three-action games}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{predicted probabilities of actions}
  \item{columns}{treatments}
  \item{tables}{different theories}
  \item{cells}{probability of choosing an action on each treatment depending on the theory}
  ...
}
}
\source{
\url{http://hdg.kuelpmann.org/}
}
\usage{
predictions_three_action_games
}
\description{
Blabla!
}
\keyword{datasets}
