% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predictions_two_action_games}
\alias{predictions_two_action_games}
\title{Predictions for the two-action games}
\format{
\describe{
  \item{rows}{predicted probabilities of actions}
  \item{columns}{treatments}
  \item{tables}{different theories}
  \item{cells}{probability of choosing an action on each treatment depending on the theory}
  ...
}
}
\source{
\url{http://hdg.kuelpmann.org/}
}
\usage{
predictions_two_action_games
}
\description{
These are the prediction of every theory considered in Kuelpmann and Kuzmics (2022) for the two-action games.
For a explanation of the different theories, treatments and the games, please refer to the paper.
}
\keyword{datasets}
