% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsp_data}
\alias{rsp_data}
\title{RSP data (3x3)}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  ...
}
}
\source{
\url{http://madeitup.myself/}
}
\usage{
rsp_data
}
\description{
Blabla!
}
\keyword{datasets}
