% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_old.R
\name{vuong_matrix_2}
\alias{vuong_matrix_2}
\title{Getting a vuong matrix}
\usage{
vuong_matrix_2(data, theories)
}
\arguments{
\item{data}{Data of the experiment (matrix of ints/data.frame): columns are observations of action 1,2,3..., rows are treatments.}

\item{theories}{Model predictions for all theories. (how to structure? List of matrices?)}
}
\value{
Table of z-scores for all comparisons
}
\description{
Getting a vuong matrix
}
\examples{
(missing)
}
