% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_statistic.R
\name{vuong_statistic}
\alias{vuong_statistic}
\title{Implementation of the Vuong test for our setting}
\usage{
vuong_statistic(data, pred_I, pred_J)
}
\arguments{
\item{data}{Data of the experiment (matrix of ints/data.frame): columns are observations of action 1,2,3..., rows are treatments.}

\item{pred_I}{Model prediction for a theory (matrix of doubles/data.frame): columns are prediction of playing action 1,2,3..., rows are treatments}

\item{pred_J}{Model prediction for a theory (matrix of doubles/data.frame): columns are prediction of playing action 1,2,3..., rows are treatments}
}
\value{
The z score of testing theory I against theory J, given the data
}
\description{
Implementation of the Vuong test for our setting
}
\examples{
(missing)
}
