% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_likelihood_table.R
\name{create_likelihood_table}
\alias{create_likelihood_table}
\title{Getting a likelihood table}
\usage{
create_likelihood_table(data, predictions)
}
\arguments{
\item{data}{Data of the experiment (matrix of ints/data.frame): columns are observations of action 1,2,3..., rows are treatments.}

\item{predictions}{Model prediction for one theory: rows/cols as data}
}
\value{
Table of z-scores for all comparisons
}
\description{
Getting a likelihood table
}
\examples{
(missing)
}
