% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_three_action_games}
\alias{data_three_action_games}
\title{Data from the three-action games}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  \item{cells}{number of subjects who chose each action on each treatment}
  ...
}
}
\source{
\url{http://hdg.kuelpmann.org/}
}
\usage{
data_three_action_games
}
\description{
This is the choice data from Kuelpmann and Kuzmics (2022) for the three-action games, i.e., Hawk-Middle-Dove and Rock-Paper-Scissor.
For an explanation of the different treatments and the games, please refer to the paper.
}
\keyword{datasets}
