% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_two_action_games}
\alias{data_two_action_games}
\title{Data from the two-action games}
\format{
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  \item{cells}{number of subjects who chose each action on each treatment}
  ...
}
}
\source{
\url{https://homepage.univie.ac.at/philipp.kuelpmann/hdg.html}
}
\usage{
data_two_action_games
}
\description{
This is the choice data from Kuelpmann and Kuzmics (2022) for the two-action games, i.e., Hawk-Dove and Matching Pennies.
For an explanation of the different treatments and the games, please refer to the paper.
}
\keyword{datasets}
