% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_two_action_games}
\alias{data_two_action_games}
\title{Data from the two-action games}
\format{
Need to reformat anyway ...
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  ...
}
}
\source{
\url{http://madeitup.myself/}
}
\usage{
data_two_action_games
}
\description{
Blabla!
}
\keyword{datasets}
