% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lh.R
\name{log_lh}
\alias{log_lh}
\title{Calculate log-likelihood for a theory}
\usage{
log_lh(model_prediction, data_experiment)
}
\arguments{
\item{model_prediction}{Model prediction for a theory (matrix of doubles)}

\item{data_experiment}{Data of the experiment (matrix of ints)}
}
\value{
The log-likelihood of the theory
}
\description{
Calculate log-likelihood for a theory
}
