% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_test_data}
\alias{simple_test_data}
\title{Simple Test Data}
\format{
A matrix with 3 rows and 2 columns
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  \item{cells}{number of subjects who chose each action on each treatment}
  ...
}
}
\usage{
simple_test_data
}
\description{
This is just very simple test data, mostly used for testing and demonstration purposes.
}
\keyword{datasets}
