% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_test_data}
\alias{simple_test_data}
\title{Simple Test Data}
\format{
A matrix with 3 rows and 2 columns
\describe{
  \item{rows}{actions}
  \item{columns}{treatments}
  ...
}
}
\source{
\url{http://madeitup.myself/}
}
\usage{
simple_test_data
}
\description{
Mostly to test data im/exporting
Don't make accessible to users?
}
\keyword{datasets}
