% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_statistic.R
\name{vuong_statistic}
\alias{vuong_statistic}
\title{Implementation of the Vuong test for our setting}
\usage{
vuong_statistic(data, pred_i, pred_j)
}
\arguments{
\item{data}{Data of the experiment (matrix of ints/data.frame): columns are treatments, rows are actions, values are number of subjects who chose each action on each treatment}

\item{pred_i}{Model prediction for a theory (matrix of doubles/data.frame): rows are prediction of playing action 1,2,3..., columns are treatments}

\item{pred_j}{Model prediction for a theory (matrix of doubles/data.frame): rows are prediction of playing action 1,2,3..., columns are treatments}
}
\value{
The z score of testing theory I against theory J, given the data
}
\description{
Implementation of the Vuong test for our setting
}
\examples{
vuong_statistic(
  data_two_action_games,
  predictions_two_action_games[, , 1],
  predictions_two_action_games[, , 2]
)
}
